work process migration issue solved

+++ src/TalentTracker.Domain/Aggregates/Container/Entities/WorkProcessIngredient.cs
@@ -8,9 +8,12 @@ namespace TalentTracker.Domain.Aggregates.Container.Entities;
 public class WorkProcessIngredient : AuditableEntity, IAggregateRoot
 {
     public Guid WorkProcessIngredientGUID { get; private set; }
-    public int WorkProcessID { get; private set; }
+    public long WorkProcessID { get; private set; }
     public int IngredientID { get; private set; }
     public MeasurementUnit MeasurementUnit { get; private set; }
+
+    //public MeasurementUnitType MeasurementUnitType { get; private set; }
+    //public string MeasurementValue { get; private set; }
     public DateTime Date { get; private set; }
     public Quantity Quantity { get; private set; }
 
@@ -20,15 +23,19 @@ public class WorkProcessIngredient : AuditableEntity, IAggregateRoot
 
     public WorkProcessIngredient()
     {
-            
+
     }
 
     // Constructor for setting properties
-    public WorkProcessIngredient(int ingredientID, MeasurementUnit measurementUnit, DateTime date, Quantity quantity)
+    public WorkProcessIngredient(int ingredientID, MeasurementUnit measurementUnit,
+        //MeasurementUnitType measurementUnitType, string measurementValue, 
+        DateTime date, Quantity quantity)
     {
         WorkProcessIngredientGUID = Guid.NewGuid();
         IngredientID = Guard.Against.NegativeOrZero(ingredientID);
         MeasurementUnit = Guard.Against.Default(measurementUnit);
+        //MeasurementUnitType = Guard.Against.Default(measurementUnitType);
+        //MeasurementValue = Guard.Against.NullOrEmpty(measurementValue);
         Date = Guard.Against.Default(date);
         Quantity = Guard.Against.Default(quantity);
     }
